Refactor NewListing into separate component
