+++ package.json
@@ -10,6 +10,8 @@
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
+    "@smui/button": "^6.0.0-beta.14",
+    "@smui/card": "^6.0.0-beta.14",
     "rollup": "^2.3.4",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-livereload": "^2.0.0",
@@ -21,6 +23,7 @@
   "dependencies": {
     "firebase": "^9.6.7",
     "sirv-cli": "^2.0.2",
+    "svelte-material-ui": "^6.0.0-beta.14",
     "svelte-navigator": "^3.1.5"
   }
 }
